pyproject.toml
[project]
name = "prezentation-feedback-agent"
version = "0.1.0"
description = "Presentation Feedback Agent using Strands Agents and Amazon Bedrock"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=1.6.0",
    "strands-agents-tools>=0.2.5",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "pydub>=0.25.1",
    "streamlit>=1.46.1",
]
